[Rename] 프로젝트 폴더 구조 수정
