+++ src/app/page.tsx
@@ -2,7 +2,7 @@ import Image from 'next/image'
 import styles from './page.module.css'
 import React from 'react';
 import { useState } from 'react';
-import Read from './read/read'
+import Read from '../pages/user/read'
 
 export default function Home() {
   return (
