+++ src/types/user/read.ts
@@ -1,0 +1,9 @@
+export interface UserData {
+    email: string;
+    username: string;
+    phone: string;
+    birth: string;
+    profile_url: string;
+    level: number;
+    sns: string[];
+}
